func: création de jury
